+++ text/message_recovery.go
@@ -23,7 +23,7 @@ func NewRecoverySuccessful(privilegedSessionExpiresAt time.Time) *Message {
 	hasLeft := Until(privilegedSessionExpiresAt)
 	return &Message{
 		ID:   InfoSelfServiceRecoverySuccessful,
-		Type: Info,
+		Type: Success,
 		Text: fmt.Sprintf("You successfully recovered your account. Please change your password or set up an alternative login method (e.g. social sign in) within the next %.2f minutes.", hasLeft.Minutes()),
 		Context: context(map[string]interface{}{
 			"privilegedSessionExpiresAt": privilegedSessionExpiresAt,